feat: change to TS files
