+++ src/component/Display.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import './Display.css'
+import { DisplayProps } from './types'
 
 // export default class Display extends React.Component {
 //   static propTypes = {
@@ -17,7 +18,7 @@ import './Display.css'
 //   }
 // }
 
-export default function Display({ value }) {
+export default function Display({ value }: DisplayProps) {
   return (
     <div className="component-display">
       <div>{value}</div>
