+++ src/logic/calculate.d.ts
@@ -1,0 +1,6 @@
+import { AppState, PartialAppState } from '../component/types'
+
+export default function calculate(
+  obj: AppState,
+  buttonName: string,
+): PartialAppState
